// MoneyMeFrontEnd/src/Models/response/get-quote-response.ts
export class GetQuoteResponse {
  customerId: number = 0;
  name: string = "";
  mobile: string = "";
  email: string = "";
  financeAmount: number = 0;
  term: number = 0;
  repaymentsFrom: number = 0;
  paymentType: string = "";
  totalRepayments: number = 0;
  establishmentFee: number = 0;
  interest: number = 0;
}